restructure src folder
