+++ src/pages/ProfileContainer.js
@@ -1,5 +1,5 @@
-import ProfileMissions from '../components/ProfileMissions';
-import ProfileRockets from '../components/ProfileRockets';
+import ProfileMissions from '../components/MissionsComponents/ProfileMissions';
+import ProfileRockets from '../components/RocketsComponents/ProfileRockets';
 
 const ProfileContainer = () => (
   <div className="profileContainer">
